+++ src/api/dm/kit/assets/fonts/font_size.d
@@ -3,7 +3,7 @@ module api.dm.kit.assets.fonts.font_size;
 /**
  * Authors: initkfs
  */
-enum FontSize : size_t
+enum FontSize : uint
 {
     large,
     medium,
